% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/las_mod_functions.R
\name{las_convert_v2<-}
\alias{las_convert_v2<-}
\title{Change the version of the LAS object to 2.0}
\usage{
las_convert_v2(x) <- value
}
\arguments{
\item{x}{LAS object}

\item{value}{Boolean TRUE or FALSE}
}
\value{
Modifies LAS object in place, changing the version to V2.0, with the appropriate switching of the parameters in the WELL table
}
\description{
\code{las_convert_v2<-} Takes a LAS object and changes the version to V2.0 and modifies the WELL table
}
