% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/las_mod_functions.R
\name{las_remove_uwi<-}
\alias{las_remove_uwi<-}
\title{Remove the UWI line in the WELL table}
\usage{
las_remove_uwi(x) <- value
}
\arguments{
\item{x}{LAS object}

\item{value}{boolean TRUE or FALSE}
}
\value{
Modifies LAS object in place, removing the UWI parameter from the WELL table
}
\description{
\code{las_remove_uwi<-} Takes a LAS object and removes the UWI line in the WELL table
}
