% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/las_mod_functions.R
\name{las_set_start_depth<-}
\alias{las_set_start_depth<-}
\title{Change the start depth of a LAS object in place}
\usage{
las_set_start_depth(x) <- value
}
\arguments{
\item{x}{LAS object}

\item{value}{Number to insert as the new start depth.}
}
\value{
Modifies LAS object in place, changing the STRT parameter of the WELL information block and removing rows from the log that are lower than this number.
This was designed to remove negative starting DEPTHS, so caution should be used if the DEPTHS are not in ascending order.
}
\description{
\code{las_set_start_depth<-} Takes a LAS object and changes the STRT depth in the WELL table of the LAS object
}
