% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/las_mod_functions.R
\name{las_trim_well_id<-}
\alias{las_trim_well_id<-}
\title{Remove the UWI line in the WELL table}
\usage{
las_trim_well_id(x) <- value
}
\arguments{
\item{x}{LAS object}

\item{value}{Boolean TRUE or FALSE}
}
\value{
Modifies LAS object in place, removing spaces from the WELL reference in the WELL table.
}
\description{
\code{las_trim_well_id<-} Takes a LAS object and removes spaces from the WELL reference in the WELL table.
}
