% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/las_mod_functions.R
\name{las_version<-}
\alias{las_version<-}
\title{Convert the version number of a LAS object}
\usage{
las_version(x) <- value
}
\arguments{
\item{x}{LAS object}

\item{value}{A number, currently needs to be 1.2 or 2.0}
}
\value{
Modifies LAS object in place, changing the LAS version to the nominated number
}
\description{
\code{las_version<-} Takes a LAS object created with the read_las() function and changes the CWLS LAS version.
lastools current supports 1.2 and 2.0 only.
}
