% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/las_read.R
\name{read_las}
\alias{read_las}
\title{Read a LAS file}
\usage{
read_las(filepath, replace_null = T)
}
\arguments{
\item{filepath}{string       path to the LAS file}

\item{replace_null}{boolean  replace the NULLS in the LOG section (as given by the WELL table) to the R NA value}
}
\value{
Returns a list object with VERSION, WELL, CURVE, PARAMETER, OTHER and LOG sections. WELL, CURVE, PARAMETER and LOG are data.frames. OTHER is a string with line breaks. VERSION is a numeric (1.2 or 2.0)
}
\description{
Reads a CWLS LAS file and returns a list object with the VERSION, WELL, CURVE, PARAMETER, OTHER and LOG sections, as well as storing the original path of the file.
}
