% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/las_read_data_dt.R
\name{read_las_data_dt}
\alias{read_las_data_dt}
\title{Read las log data to data table}
\usage{
read_las_data_dt(dir)
}
\arguments{
\item{dir}{string directory path containing .las file/s}
}
\value{
Returns a long format data.table containing well_name, DEPT (Depth),variable,value and file (location of las file)
}
\description{
Reads CWLS LAS file/s from a directory location and returns a data.table of the log section data by well
}
