% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/las_write.R
\name{write_las}
\alias{write_las}
\title{Write LAS object to LAS file}
\usage{
write_las(las, path)
}
\arguments{
\item{las}{In memory las object}

\item{path}{Directory path to write las file}
}
\value{
Writes a las object as a LAS file to disk
}
\description{
Reads CWLS LAS file/s from a directory location and returns a data.table of the log section data by well
}
